New Tokenizer object.
The previous code was using the tokenizer within the `WMT2014Dataset`
object. This was problematic because we have to use tokenization for
other things (e.g., translating text from the Transformer) but can't if
it's coupled with the dataset.

It also doesn't make that much sense semantically since a dataset
doesn't need to have its own tokenizer.