Removed convert_ids_to_tokens and added decode_ids.
Previously the `convert_ids_to_tokens` was used to translate IDs intp
texts. The name was incorrect because we weren't technically converting
the IDs to their respective tokens, but we were actually decoding the
IDs to the appropriate text.

This method was therefore removed and the `decode_ids` method was added
to take care of that functionality.