Modified attention modules to receive masks as arg.
Previously, the attention function would receive a `mask` argument as a
Boolean flag, telling the function to create an upper triangular causals
mask. However, now that we want to be able to handle different types of
masks (i.e., padding mask, causal mask), it's better to just create the
mask outside of the function/modules and feed them as arguments.